requirements.txt
opencv-python
torch
torchvision
ultralytics
numpy
pillow
simpleaudio
